Drop REF_CASE_INCOMPATIBLE (Jsonnet now supports ref_declarations on
main) and keep CASE_LANGUAGE_INCOMPATIBLE for the SML/op incompatibility.

+++ tests/integration/cases/scalar_int_large/V_combined.v
@@ -1,6 +1,6 @@
 
 fn main() {
-	mut my_data := 2147483648
-	my_data = 2147483648
+	mut my_data := i64(2147483648)
+	my_data = i64(2147483648)
 	_ = my_data
 }
