+++ tests/integration/cases/scalar_int_large/V_numeric_separator_underscore.v
@@ -1,5 +1,5 @@
 
 fn main() {
-	my_data := 2_147_483_648
+	my_data := i64(2_147_483_648)
 	_ = my_data
 }
